fix testimonials, job title cropped at the bottom
The size of the items is recalculated by the javascript on every window resize. Currently the job title was cropped on some devices and not properly recalculated. This is just a workaround adding more whitespace. To fit the calculations you need to better understand how the JS gets the size and how the inserted divs need to have better marked up with padding/margin.